Moved a substantial amount (more) of the menu code into libsqee. TypeDefs for shared pointers of sq::Handler and sq::Scene. Alot of code is waaay nicer now. Replaced a few floats with doubles, and tidied a few calculations. VERY slight performance and accuracy improvements.
Also did a bit more work on the Flappy demo, but it's still not a game yet.

+++ tests/soko/scenegame.hpp
@@ -11,7 +11,7 @@ public:
 
     bool active;
     std::unique_ptr<Level> level;
-    short int pFace, pDir, pX, pY, pMoved;
+    int pFace, pDir, pX, pY, pMoved;
     float pSpeed;
     sf::Vector2i size;
     sf::Vector2i pushing;
